增加功能：上电开机未连接蓝牙也可以按键关机。重构：分板子目录，删除大部分extern
