+++ app/project/proj.conf
@@ -52,6 +52,7 @@ CONFIG_CPU_USAGE_METRICS_ENABLED=y
 CONFIG_CPU_USAGE_METRICS_PRINT_DIRECTLY=y
 CONFIG_AUDIO=y
 CONFIG_AUDIO_LOCAL_MUSIC=y
+CONFIG_PKG_USING_ANYKA=y
 CONFIG_MULTI_CLIENTS_AT_WORKING=y
 CONFIG_PKG_USING_CJSON=y
 CONFIG_BSP_USING_CMSIS_DSP=y
