+++ app/src/kws/app_recorder_process.h
@@ -1,0 +1,10 @@
+#ifndef __KWS_APP_RECORDER_PROCESS_H
+#define __KWS_APP_RECORDER_PROCESS_H
+
+#include <stdio.h>
+#include <stdint.h>
+
+void kws_demo();
+void kws_demo_stop();
+
+#endif
