+++ app/src/main.c
@@ -21,30 +21,15 @@
 #include "xiaozhi_weather.h"
 #include "lv_timer.h"
 #include "lv_display.h"
-extern void xiaozhi_ui_update_ble(char *string);
-extern void xiaozhi_ui_update_emoji(char *string);
-extern void xiaozhi_ui_chat_status(char *string);
-extern void xiaozhi_ui_chat_output(char *string);
-extern void xiaozhi_ui_standby_chat_output(char *string);
-extern void ui_swith_to_standby_screen();
-extern void ui_swith_to_xiaozhi_screen();
-extern void xiaozhi_ui_task(void *args);
-extern void xiaozhi(int argc, char **argv);
-extern void xiaozhi2(int argc, char **argv);
-extern void reconnect_xiaozhi();
-extern void xz_ws_button_init(void);
-extern void xz_ws_audio_init();
-extern void xz_mqtt_button_init(void);
-extern void xz_audio_init();
-extern rt_tick_t last_listen_tick;
-extern xiaozhi_ws_t g_xz_ws;
-extern rt_mailbox_t g_button_event_mb;
-extern void ui_sleep_callback(lv_timer_t *timer);
-extern lv_obj_t *standby_screen;
-rt_mailbox_t g_battery_mb;
-extern lv_timer_t *ui_sleep_timer;
-extern lv_obj_t *shutdown_screen;
-extern lv_obj_t *sleep_screen;
+#include "../board/board_hardware.h"
+#include "xiaozhi_ui.h"
+#include "xiaozhi.h"
+#include "xiaozhi_audio.h"
+#include "bts2_app_inc.h"
+#include "ble_connection_manager.h"
+#include "bt_connection_manager.h"
+#include "bt_env.h"
+#include "ulog.h"
 /* Common functions for RT-Thread based platform
  * -----------------------------------------------*/
 /**
@@ -52,22 +37,8 @@ extern lv_obj_t *sleep_screen;
  * @param  None
  * @retval None
  */
-void HAL_MspInit(void)
-{
-    //__asm("B .");        /*For debugging purpose*/
-    BSP_IO_Init();
-#ifdef BSP_USING_BOARD_SF32LB52_XTY_AI
-    HAL_PIN_Set(PAD_PA38, GPTIM1_CH1, PIN_PULLUP, 1);
-    HAL_PIN_Set(PAD_PA40, GPTIM1_CH2, PIN_PULLUP, 1);
-#endif
-}
 /* User code start from here
  * --------------------------------------------------------*/
-#include "bts2_app_inc.h"
-#include "ble_connection_manager.h"
-#include "bt_connection_manager.h"
-#include "bt_env.h"
-#include "ulog.h"
 
 #define BT_APP_READY 0
 #define BT_APP_CONNECT_PAN 1
@@ -82,24 +53,32 @@ void HAL_MspInit(void)
 #define UPDATE_REAL_WEATHER_AND_TIME 11
 #define PAN_TIMER_MS 3000
 #define MQTT_RECONNECT 12
+#define MAX_RECONNECT_ATTEMPTS 30  // 30次尝试，每次1秒，共30秒
+#define XIAOZHI_UI_THREAD_STACK_SIZE (6144)
+#define BATTERY_THREAD_STACK_SIZE (2048)
+
+
+extern rt_tick_t last_listen_tick;
+extern xiaozhi_ws_t g_xz_ws;
+extern rt_mailbox_t g_button_event_mb;
+extern lv_obj_t *standby_screen;
+extern lv_timer_t *ui_sleep_timer;
+extern lv_obj_t *shutdown_screen;
+extern lv_obj_t *sleep_screen;
 
 bt_app_t g_bt_app_env;
 rt_mailbox_t g_bt_app_mb;
 BOOL g_pan_connected = FALSE;
 BOOL first_pan_connected = FALSE;
 int first_reconnect_attempts = 0;
+uint8_t Initiate_disconnection_flag = 0;//蓝牙主动断开标志
+rt_mailbox_t g_battery_mb;
 
 static rt_timer_t s_reconnect_timer = NULL;
 static rt_timer_t s_sleep_timer = NULL;
 static int reconnect_attempts = 0;
-#define MAX_RECONNECT_ATTEMPTS 30  // 30次尝试，每次1秒，共30秒
 static uint8_t g_sleep_enter_flag = 0;    // 进入睡眠标志位
-uint8_t Initiate_disconnection_flag = 0;//蓝牙主动断开标志
-
-
-
-#define XIAOZHI_UI_THREAD_STACK_SIZE (6144)
-#define BATTERY_THREAD_STACK_SIZE (2048)
+// UI线程和battery线程控制块
 static struct rt_thread xiaozhi_ui_thread;
 static struct rt_thread battery_thread;
 //ui线程
@@ -195,7 +174,12 @@ static void pulse_encoder_timeout_handle(void *parameter)
     }
 }
 #endif
-
+void HAL_MspInit(void)
+{
+    //__asm("B .");        /*For debugging purpose*/
+    BSP_IO_Init();
+    set_pinmux();
+}
 static void battery_level_task(void *parameter)
 {
     g_battery_mb = rt_mb_create("battery_level", 1, RT_IPC_FLAG_FIFO);
@@ -556,64 +540,6 @@ uint32_t bt_get_class_of_device()
            BT_PERIPHERAL_REMCONTROL;
 }
 
-static void check_poweron_reason(void)
-{
-    switch (SystemPowerOnModeGet())
-    {
-    case PM_REBOOT_BOOT:
-    case PM_COLD_BOOT:
-    {
-        // power on as normal
-        break;
-    }
-    case PM_HIBERNATE_BOOT:
-    case PM_SHUTDOWN_BOOT:
-    {
-        if (PMUC_WSR_RTC & pm_get_wakeup_src())
-        {
-            // RTC唤醒
-            NVIC_EnableIRQ(RTC_IRQn);
-            // power on as normal
-        }
-#ifdef BSP_USING_CHARGER
-        else if ((PMUC_WSR_PIN0 << (pm_get_charger_pin_wakeup())) & pm_get_wakeup_src())
-        {
-        }
-#endif
-        else if (PMUC_WSR_PIN_ALL & pm_get_wakeup_src())
-        {
-            rt_thread_mdelay(1000); // 延时1秒
-#ifdef BSP_USING_BOARD_SF32LB52_LCD_N16R8
-            int val = rt_pin_read(BSP_KEY1_PIN);
-#else
-            int val = rt_pin_read(BSP_KEY2_PIN);
-#endif
-            rt_kprintf("Power key level after 1s: %d\n", val);
-            if (val != KEY2_ACTIVE_LEVEL)
-            {
-                // 按键已松开，认为是误触发，直接关机
-                rt_kprintf("Not long press, shutdown now.\n");
-                PowerDownCustom();
-                while (1) {};
-            }
-            else
-            {
-                // 长按，正常开机
-                rt_kprintf("Long press detected, power on as normal.\n");
-            }
-        }
-        else if (0 == pm_get_wakeup_src())
-        {
-            RT_ASSERT(0);
-        }
-        break;
-    }
-    default:
-    {
-        RT_ASSERT(0);
-    }
-    }
-}
 static int32_t Write_MAC(int argc, char **argv)
 {
     uint8_t len;
@@ -667,22 +593,9 @@ int main(void)
 #ifdef XIAOZHI_USING_MQTT
 #else
     xz_ws_audio_init(); // 初始化音频
+    xz_ws_button_init2();//初始化关机键
 #endif
-
-#ifdef BSP_USING_BOARD_SF32LB52_LCHSPI_ULP
-    unsigned int *addr2 = (unsigned int *)0x50003088; // 21
-    *addr2 = 0x00000200;
-    unsigned int *addr = (unsigned int *)0x500030B0; // 31
-    *addr = 0x00000200;
-
-    // senser
-    HAL_PIN_Set(PAD_PA30, GPIO_A30, PIN_PULLDOWN, 1);
-    BSP_GPIO_Set(30, 0, 1);
-    HAL_PIN_Set(PAD_PA39, GPIO_A39, PIN_PULLDOWN, 1);
-    HAL_PIN_Set(PAD_PA40, GPIO_A40, PIN_PULLDOWN, 1);
-
-    // rt_pm_request(PM_SLEEP_MODE_IDLE);
-#endif
+    set_pinmux();
     // Create  xiaozhi UI
     rt_err_t result = rt_thread_init(&xiaozhi_ui_thread,
                                      "xz_ui",
