+++ app/src/mcp/mcp_api.cc
@@ -4,7 +4,7 @@
 #include <rtthread.h>
 #include <string.h>
 #include "rgbled_mcp.h"
-
+#include "../board/board_hardware.h"
 
 extern "C" 
 {
@@ -16,18 +16,8 @@ extern "C"
 
     void MCP_RGBLED_CLOSE()
     {
-#if defined(BSP_USING_BOARD_SF32LB52_NANO_52J) || defined(BSP_USING_BOARD_SF32LB52_XTY_AI_THT)
-        rt_pin_write(32, PIN_HIGH);
-#else
-        if (RGBLEDTool::is_color_cycling_)
-        {
-            RGBLEDTool::is_color_cycling_ = false;
-            rt_thread_mdelay(100);
-            GetRGBLEDController().SetColor(0x000000);  // 关灯
-        }
-#endif
-
-    }  
+        close_led();
+    }
