+++ app/src/mcp/rgbled_mcp.cc
@@ -5,7 +5,7 @@
 #include "string.h"
 #include "drivers/rt_drv_pwm.h"
 #include "bf0_hal.h"
-
+#include "../board/board_hardware.h" 
 #define LED_PIN 32 
 RGBLEDController& GetRGBLEDController() {
     static RGBLEDController instance;
@@ -13,22 +13,9 @@ RGBLEDController& GetRGBLEDController() {
 }
 
 bool RGBLEDTool::is_color_cycling_ = false;
-void RGBLEDTool::ColorCycleThreadEntry(void* param) {
-    int index = 0;
-    uint32_t toggle = 0; // 用于52J板子的LED切换状态
-    while (is_color_cycling_) {
-#if defined(BSP_USING_BOARD_SF32LB52_NANO_52J) || defined(BSP_USING_BOARD_SF32LB52_XTY_AI_THT)
-        // 对于52J板子，控制PA32引脚以1秒周期亮灭
-        rt_pin_write(LED_PIN, toggle ? PIN_HIGH : PIN_LOW);
-        toggle = !toggle;
-        rt_thread_mdelay(1000);
-#else
-        uint32_t color = rgb_color_arry[index].color;
-        GetRGBLEDController().SetColor(color);
-        index = (index + 1) % (sizeof(rgb_color_arry)/sizeof(rgb_color_arry[0]));
-        rt_thread_mdelay(500);
-#endif
-    }
+void RGBLEDTool::ColorCycleThreadEntry(void* param) 
+{
+    open_led();    
 }
 bool RGBLEDTool::IsLightOn() {
     // 如果正在循环变色，则认为灯是开启状态
@@ -43,11 +30,8 @@ void RGBLEDTool::RegisterRGBLEDTool(McpServer* server) {
             [](const PropertyList&) -> ReturnValue {
             if (is_color_cycling_) return true;
             is_color_cycling_ = true;
-#if defined(BSP_USING_BOARD_SF32LB52_NANO_52J) || defined(BSP_USING_BOARD_SF32LB52_XTY_AI_THT)
             // 配置PA32为GPIO输出模式并输出低电平（点亮）
-            rt_pin_mode(LED_PIN, PIN_MODE_OUTPUT);
-            rt_pin_write(LED_PIN, PIN_LOW);
-#endif      
+            set_mode_led(); 
             rt_thread_t thread = rt_thread_create("rgb_cycle",
                             ColorCycleThreadEntry, 
                             nullptr,
@@ -64,15 +48,7 @@ void RGBLEDTool::RegisterRGBLEDTool(McpServer* server) {
         "turn off the light.",
         PropertyList(),
         [](const PropertyList&) -> ReturnValue {
-            is_color_cycling_ = false;
-            rt_thread_mdelay(100); 
-#if defined(BSP_USING_BOARD_SF32LB52_NANO_52J) || defined(BSP_USING_BOARD_SF32LB52_XTY_AI_THT)
-            // 对于52J板子，控制PA32引脚输出高电平熄灭LED
-            rt_pin_write(LED_PIN, PIN_HIGH);
-#else
-            
-            GetRGBLEDController().SetColor(0x000000);
-#endif
+            close_led();
             return true;
         }
     );
