+++ app/src/xiaozhi_ui.h
@@ -2,6 +2,43 @@
 #define _XIAOZHI_UI_H_
 
 #include "lv_image_dsc.h"
+#include "lv_timer.h"
+#include "lv_display.h"
+#include "lv_obj_pos.h"
+
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void xiaozhi_ui_update_ble(char *string); // ble
+void xiaozhi_ui_update_emoji(char *string);
+void xiaozhi_ui_chat_status(char *string);
+void xiaozhi_ui_chat_output(char *string);
+void xiaozhi_ui_standby_chat_output(char *string);
+void ui_swith_to_standby_screen(void);
+void ui_swith_to_xiaozhi_screen(void);
+void xiaozhi_ui_task(void *args);
+void ui_sleep_callback(lv_timer_t *timer);
+void ui_swith_to_xiaozhi_screen(void);
+void xiaozhi_ui_tts_output(char *string);
+void xiaozhi_ui_update_standby_emoji(char *string);
+void show_shutdown(void);
+void show_sleep_countdown_and_sleep(void);
+
+extern const lv_image_dsc_t * angry[57];
+extern const lv_image_dsc_t * loving[48];
+extern const lv_image_dsc_t * kissy[57];
+extern const lv_image_dsc_t * cool[60];
+extern const lv_image_dsc_t * crying[26];
+extern const lv_image_dsc_t * embarrassed[57];
+extern const lv_image_dsc_t * funny[57];
+extern const lv_image_dsc_t * neutral[57];
+extern const lv_image_dsc_t * sleepy[57];
+extern const lv_image_dsc_t * surprised[57];
+extern const lv_image_dsc_t * thinking[57];
+
+
 
 
 
@@ -63,66 +100,7 @@ extern const lv_image_dsc_t angry_55;
 extern const lv_image_dsc_t angry_56;
 extern const lv_image_dsc_t angry_57;
 
-const lv_image_dsc_t * angry[] =
-{
-  &angry_1,
-  &angry_2,
-  &angry_3,
-  &angry_4,
-  &angry_5,
-  &angry_6,
-  &angry_7,
-  &angry_8,
-  &angry_9,
-  &angry_10,
-  &angry_11,
-  &angry_12,
-  &angry_13,
-  &angry_14,
-  &angry_15,
-  &angry_16,
-  &angry_17,
-  &angry_18,
-  &angry_19,
-  &angry_20,
-  &angry_21,
-  &angry_22,
-  &angry_23,
-  &angry_24,
-  &angry_25,
-  &angry_26,
-  &angry_27,
-  &angry_28,
-  &angry_29,
-  &angry_30,
-  &angry_31,
-  &angry_32,
-  &angry_33,
-  &angry_34,
-  &angry_35,
-  &angry_36,
-  &angry_37,
-  &angry_38,
-  &angry_39,
-  &angry_40,
-  &angry_41,
-  &angry_42,
-  &angry_43,
-  &angry_44,
-  &angry_45,
-  &angry_46,
-  &angry_47,
-  &angry_48,
-  &angry_49,
-  &angry_50,
-  &angry_51,
-  &angry_52,
-  &angry_53,
-  &angry_54,
-  &angry_55,
-  &angry_56,
-  &angry_57
-};
+
 /*---------------------angry*/
 
 extern const lv_image_dsc_t cool_1;
@@ -188,69 +166,7 @@ extern const lv_image_dsc_t cool_60;
 
 
 
-const lv_image_dsc_t * cool[] =
-{
-  &cool_1,
-  &cool_2,
-  &cool_3,
-  &cool_4,
-  &cool_5,
-  &cool_6,
-  &cool_7,
-  &cool_8,
-  &cool_9,
-  &cool_10,
-  &cool_11,
-  &cool_12,
-  &cool_13,
-  &cool_14,
-  &cool_15,
-  &cool_16,
-  &cool_17,
-  &cool_18,
-  &cool_19,
-  &cool_20,
-  &cool_21,
-  &cool_22,
-  &cool_23,
-  &cool_24,
-  &cool_25,
-  &cool_26,
-  &cool_27,
-  &cool_28,
-  &cool_29,
-  &cool_30,
-  &cool_31,
-  &cool_32,
-  &cool_33,
-  &cool_34,
-  &cool_35,
-  &cool_36,
-  &cool_37,
-  &cool_38,
-  &cool_39,
-  &cool_40,
-  &cool_41,
-  &cool_42,
-  &cool_43,
-  &cool_44,
-  &cool_45,
-  &cool_46,
-  &cool_47,
-  &cool_48,
-  &cool_49,
-  &cool_50,
-  &cool_51,
-  &cool_52,
-  &cool_53,
-  &cool_54,
-  &cool_55,
-  &cool_56,
-  &cool_57,
-  &cool_58,
-  &cool_59,
-  &cool_60
-};
+
 
 /*---------------------cool*/
 
@@ -282,35 +198,7 @@ extern const lv_image_dsc_t crying_24;
 extern const lv_image_dsc_t crying_25;
 extern const lv_image_dsc_t crying_26;
 
-const lv_image_dsc_t * crying[] =
-{
-  &crying_1,
-  &crying_2,
-  &crying_3,
-  &crying_4,
-  &crying_5,
-  &crying_6,
-  &crying_7,
-  &crying_8,
-  &crying_9,
-  &crying_10,
-  &crying_11,
-  &crying_12,
-  &crying_13,
-  &crying_14,
-  &crying_15,
-  &crying_16,
-  &crying_17,
-  &crying_18,
-  &crying_19,
-  &crying_20,
-  &crying_21,
-  &crying_22,
-  &crying_23,
-  &crying_24,
-  &crying_25,
-  &crying_26
-};
+
 /*---------------------crying*/
 
 
@@ -375,66 +263,7 @@ extern const lv_image_dsc_t embarrassed_57;
 
 
 
-const lv_image_dsc_t * embarrassed[] =
-{
-  &embarrassed_1,
-  &embarrassed_2,
-  &embarrassed_3,
-  &embarrassed_4,
-  &embarrassed_5,
-  &embarrassed_6,
-  &embarrassed_7,
-  &embarrassed_8,
-  &embarrassed_9,
-  &embarrassed_10,
-  &embarrassed_11,
-  &embarrassed_12,
-  &embarrassed_13,
-  &embarrassed_14,
-  &embarrassed_15,
-  &embarrassed_16,
-  &embarrassed_17,
-  &embarrassed_18,
-  &embarrassed_19,
-  &embarrassed_20,
-  &embarrassed_21,
-  &embarrassed_22,
-  &embarrassed_23,
-  &embarrassed_24,
-  &embarrassed_25,
-  &embarrassed_26,
-  &embarrassed_27,
-  &embarrassed_28,
-  &embarrassed_29,
-  &embarrassed_30,
-  &embarrassed_31,
-  &embarrassed_32,
-  &embarrassed_33,
-  &embarrassed_34,
-  &embarrassed_35,
-  &embarrassed_36,
-  &embarrassed_37,
-  &embarrassed_38,
-  &embarrassed_39,
-  &embarrassed_40,
-  &embarrassed_41,
-  &embarrassed_42,
-  &embarrassed_43,
-  &embarrassed_44,
-  &embarrassed_45,
-  &embarrassed_46,
-  &embarrassed_47,
-  &embarrassed_48,
-  &embarrassed_49,
-  &embarrassed_50,
-  &embarrassed_51,
-  &embarrassed_52,
-  &embarrassed_53,
-  &embarrassed_54,
-  &embarrassed_55,
-  &embarrassed_56,
-  &embarrassed_57
-};
+
 
 
 /*---------------------embarrassed*/
@@ -500,66 +329,6 @@ extern const lv_image_dsc_t funny_57;
 
 
 
-const lv_image_dsc_t * funny[] =
-{
-  &funny_1,
-  &funny_2,
-  &funny_3,
-  &funny_4,
-  &funny_5,
-  &funny_6,
-  &funny_7,
-  &funny_8,
-  &funny_9,
-  &funny_10,
-  &funny_11,
-  &funny_12,
-  &funny_13,
-  &funny_14,
-  &funny_15,
-  &funny_16,
-  &funny_17,
-  &funny_18,
-  &funny_19,
-  &funny_20,
-  &funny_21,
-  &funny_22,
-  &funny_23,
-  &funny_24,
-  &funny_25,
-  &funny_26,
-  &funny_27,
-  &funny_28,
-  &funny_29,
-  &funny_30,
-  &funny_31,
-  &funny_32,
-  &funny_33,
-  &funny_34,
-  &funny_35,
-  &funny_36,
-  &funny_37,
-  &funny_38,
-  &funny_39,
-  &funny_40,
-  &funny_41,
-  &funny_42,
-  &funny_43,
-  &funny_44,
-  &funny_45,
-  &funny_46,
-  &funny_47,
-  &funny_48,
-  &funny_49,
-  &funny_50,
-  &funny_51,
-  &funny_52,
-  &funny_53,
-  &funny_54,
-  &funny_55,
-  &funny_56,
-  &funny_57
-};
 
 /*---------------------funny*/
 
@@ -626,66 +395,7 @@ extern const lv_image_dsc_t kissy_57;
 
 
 
-const lv_image_dsc_t * kissy[] =
-{
-  &kissy_1,
-  &kissy_2,
-  &kissy_3,
-  &kissy_4,
-  &kissy_5,
-  &kissy_6,
-  &kissy_7,
-  &kissy_8,
-  &kissy_9,
-  &kissy_10,
-  &kissy_11,
-  &kissy_12,
-  &kissy_13,
-  &kissy_14,
-  &kissy_15,
-  &kissy_16,
-  &kissy_17,
-  &kissy_18,
-  &kissy_19,
-  &kissy_20,
-  &kissy_21,
-  &kissy_22,
-  &kissy_23,
-  &kissy_24,
-  &kissy_25,
-  &kissy_26,
-  &kissy_27,
-  &kissy_28,
-  &kissy_29,
-  &kissy_30,
-  &kissy_31,
-  &kissy_32,
-  &kissy_33,
-  &kissy_34,
-  &kissy_35,
-  &kissy_36,
-  &kissy_37,
-  &kissy_38,
-  &kissy_39,
-  &kissy_40,
-  &kissy_41,
-  &kissy_42,
-  &kissy_43,
-  &kissy_44,
-  &kissy_45,
-  &kissy_46,
-  &kissy_47,
-  &kissy_48,
-  &kissy_49,
-  &kissy_50,
-  &kissy_51,
-  &kissy_52,
-  &kissy_53,
-  &kissy_54,
-  &kissy_55,
-  &kissy_56,
-  &kissy_57
-};
+
 
 
 /*---------------------kissy*/
@@ -744,57 +454,7 @@ extern const lv_image_dsc_t loving_48;
 
 
 
-const lv_image_dsc_t * loving[] =
-{
-  &loving_1,
-  &loving_2,
-  &loving_3,
-  &loving_4,
-  &loving_5,
-  &loving_6,
-  &loving_7,
-  &loving_8,
-  &loving_9,
-  &loving_10,
-  &loving_11,
-  &loving_12,
-  &loving_13,
-  &loving_14,
-  &loving_15,
-  &loving_16,
-  &loving_17,
-  &loving_18,
-  &loving_19,
-  &loving_20,
-  &loving_21,
-  &loving_22,
-  &loving_23,
-  &loving_24,
-  &loving_25,
-  &loving_26,
-  &loving_27,
-  &loving_28,
-  &loving_29,
-  &loving_30,
-  &loving_31,
-  &loving_32,
-  &loving_33,
-  &loving_34,
-  &loving_35,
-  &loving_36,
-  &loving_37,
-  &loving_38,
-  &loving_39,
-  &loving_40,
-  &loving_41,
-  &loving_42,
-  &loving_43,
-  &loving_44,
-  &loving_45,
-  &loving_46,
-  &loving_47,
-  &loving_48
-};
+
 
 /*-----------------loving*/
 
@@ -859,66 +519,7 @@ extern const lv_image_dsc_t neutral_57;
 
 
 
-const lv_image_dsc_t * neutral[] =
-{
-  &neutral_1,
-  &neutral_2,
-  &neutral_3,
-  &neutral_4,
-  &neutral_5,
-  &neutral_6,
-  &neutral_7,
-  &neutral_8,
-  &neutral_9,
-  &neutral_10,
-  &neutral_11,
-  &neutral_12,
-  &neutral_13,
-  &neutral_14,
-  &neutral_15,
-  &neutral_16,
-  &neutral_17,
-  &neutral_18,
-  &neutral_19,
-  &neutral_20,
-  &neutral_21,
-  &neutral_22,
-  &neutral_23,
-  &neutral_24,
-  &neutral_25,
-  &neutral_26,
-  &neutral_27,
-  &neutral_28,
-  &neutral_29,
-  &neutral_30,
-  &neutral_31,
-  &neutral_32,
-  &neutral_33,
-  &neutral_34,
-  &neutral_35,
-  &neutral_36,
-  &neutral_37,
-  &neutral_38,
-  &neutral_39,
-  &neutral_40,
-  &neutral_41,
-  &neutral_42,
-  &neutral_43,
-  &neutral_44,
-  &neutral_45,
-  &neutral_46,
-  &neutral_47,
-  &neutral_48,
-  &neutral_49,
-  &neutral_50,
-  &neutral_51,
-  &neutral_52,
-  &neutral_53,
-  &neutral_54,
-  &neutral_55,
-  &neutral_56,
-  &neutral_57
-};
+
 /*---------------------neutra*/
 
 
@@ -983,66 +584,7 @@ extern const lv_image_dsc_t sleepy_57;
 
 
 
-const lv_image_dsc_t * sleepy[] =
-{
-  &sleepy_1,
-  &sleepy_2,
-  &sleepy_3,
-  &sleepy_4,
-  &sleepy_5,
-  &sleepy_6,
-  &sleepy_7,
-  &sleepy_8,
-  &sleepy_9,
-  &sleepy_10,
-  &sleepy_11,
-  &sleepy_12,
-  &sleepy_13,
-  &sleepy_14,
-  &sleepy_15,
-  &sleepy_16,
-  &sleepy_17,
-  &sleepy_18,
-  &sleepy_19,
-  &sleepy_20,
-  &sleepy_21,
-  &sleepy_22,
-  &sleepy_23,
-  &sleepy_24,
-  &sleepy_25,
-  &sleepy_26,
-  &sleepy_27,
-  &sleepy_28,
-  &sleepy_29,
-  &sleepy_30,
-  &sleepy_31,
-  &sleepy_32,
-  &sleepy_33,
-  &sleepy_34,
-  &sleepy_35,
-  &sleepy_36,
-  &sleepy_37,
-  &sleepy_38,
-  &sleepy_39,
-  &sleepy_40,
-  &sleepy_41,
-  &sleepy_42,
-  &sleepy_43,
-  &sleepy_44,
-  &sleepy_45,
-  &sleepy_46,
-  &sleepy_47,
-  &sleepy_48,
-  &sleepy_49,
-  &sleepy_50,
-  &sleepy_51,
-  &sleepy_52,
-  &sleepy_53,
-  &sleepy_54,
-  &sleepy_55,
-  &sleepy_56,
-  &sleepy_57
-};
+
 
 /*---------------------sleepy*/
 
@@ -1108,66 +650,7 @@ extern const lv_image_dsc_t surprised_57;
 
 
 
-const lv_image_dsc_t * surprised[] =
-{
-  &surprised_1,
-  &surprised_2,
-  &surprised_3,
-  &surprised_4,
-  &surprised_5,
-  &surprised_6,
-  &surprised_7,
-  &surprised_8,
-  &surprised_9,
-  &surprised_10,
-  &surprised_11,
-  &surprised_12,
-  &surprised_13,
-  &surprised_14,
-  &surprised_15,
-  &surprised_16,
-  &surprised_17,
-  &surprised_18,
-  &surprised_19,
-  &surprised_20,
-  &surprised_21,
-  &surprised_22,
-  &surprised_23,
-  &surprised_24,
-  &surprised_25,
-  &surprised_26,
-  &surprised_27,
-  &surprised_28,
-  &surprised_29,
-  &surprised_30,
-  &surprised_31,
-  &surprised_32,
-  &surprised_33,
-  &surprised_34,
-  &surprised_35,
-  &surprised_36,
-  &surprised_37,
-  &surprised_38,
-  &surprised_39,
-  &surprised_40,
-  &surprised_41,
-  &surprised_42,
-  &surprised_43,
-  &surprised_44,
-  &surprised_45,
-  &surprised_46,
-  &surprised_47,
-  &surprised_48,
-  &surprised_49,
-  &surprised_50,
-  &surprised_51,
-  &surprised_52,
-  &surprised_53,
-  &surprised_54,
-  &surprised_55,
-  &surprised_56,
-  &surprised_57
-};
+
 
 /*---------------------surprised*/
 
@@ -1232,70 +715,10 @@ extern const lv_image_dsc_t thinking_57;
 
 
 
-
-const lv_image_dsc_t * thinking[] =
-{
-  &thinking_1,
-  &thinking_2,
-  &thinking_3,
-  &thinking_4,
-  &thinking_5,
-  &thinking_6,
-  &thinking_7,
-  &thinking_8,
-  &thinking_9,
-  &thinking_10,
-  &thinking_11,
-  &thinking_12,
-  &thinking_13,
-  &thinking_14,
-  &thinking_15,
-  &thinking_16,
-  &thinking_17,
-  &thinking_18,
-  &thinking_19,
-  &thinking_20,
-  &thinking_21,
-  &thinking_22,
-  &thinking_23,
-  &thinking_24,
-  &thinking_25,
-  &thinking_26,
-  &thinking_27,
-  &thinking_28,
-  &thinking_29,
-  &thinking_30,
-  &thinking_31,
-  &thinking_32,
-  &thinking_33,
-  &thinking_34,
-  &thinking_35,
-  &thinking_36,
-  &thinking_37,
-  &thinking_38,
-  &thinking_39,
-  &thinking_40,
-  &thinking_41,
-  &thinking_42,
-  &thinking_43,
-  &thinking_44,
-  &thinking_45,
-  &thinking_46,
-  &thinking_47,
-  &thinking_48,
-  &thinking_49,
-  &thinking_50,
-  &thinking_51,
-  &thinking_52,
-  &thinking_53,
-  &thinking_54,
-  &thinking_55,
-  &thinking_56,
-  &thinking_57
-};
-
 /*---------------------thinking*/
 
-
+#ifdef __cplusplus
+}
+#endif
 
 #endif // _XIAOZHI_UI_H_
